# server/data/openapi_schema.json
{"openapi": "3.0.2", "info": {"title": "Custom title", "description": "This is a very custom OpenAPI schema", "version": "2.5.0"}, "paths": {"/luncho": {"get": {"tags": ["Luncho"], "summary": "Luncho", "operationId": "luncho", "parameters": [{"required": false, "schema": {"title": "Country Code", "type": "string", "default": "JPN"}, "name": "country_code", "in": "query"}, {"required": false, "schema": {"title": "Luncho Value", "type": "number", "default": 100}, "name": "luncho_value", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LunchoResult"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/lunchos": {"get": {"tags": ["Luncho"], "summary": "Lunchos", "operationId": "lunchos", "parameters": [{"required": true, "schema": {"title": "Luncho Value", "type": "number"}, "name": "luncho_value", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Lunchos Lunchos Get", "type": "array", "items": {"$ref": "#/components/schemas/LunchoResult"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/countries": {"get": {"tags": ["Luncho"], "summary": "Countries", "description": "Returns country data for all countries.", "operationId": "countries", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Countries Countries Get", "type": "object", "additionalProperties": {"$ref": "#/components/schemas/IMF_PPP_Country"}}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "IMF_PPP_Country": {"title": "IMF_PPP_Country", "type": "object", "properties": {"year_ppp": {"title": "Year Ppp", "type": "object", "additionalProperties": {"type": "number"}}, "currency_code": {"title": "Currency Code", "type": "string"}, "currency_name": {"title": "Currency Name", "type": "string"}, "country_name": {"title": "Country Name", "type": "string"}}}, "LunchoResult": {"title": "LunchoResult", "required": ["dollar_value", "local_currency_value", "currency_code", "country_code", "country_name", "currency_name", "dollar_per_luncho"], "type": "object", "properties": {"dollar_value": {"title": "Dollar Value", "type": "number"}, "local_currency_value": {"title": "Local Currency Value", "type": "number"}, "currency_code": {"title": "Currency Code", "type": "string"}, "country_code": {"title": "Country Code", "type": "string"}, "country_name": {"title": "Country Name", "type": "string"}, "currency_name": {"title": "Currency Name", "type": "string"}, "ppp": {"title": "Ppp", "type": "number"}, "dollar_per_luncho": {"title": "Dollar Per Luncho", "type": "number"}, "exchange_rate": {"title": "Exchange Rate", "type": "number"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}